+++ packages/colony-js-adapter-ethers/src/EthersAdapter.js
@@ -54,11 +54,12 @@ export default class EthersAdapter implements IAdapter {
     this.wallet = wallet;
   }
   async getContract(query: Query): Promise<IContract> {
-    const { address, abi } = await this.loader.load(query, {
-      abi: true,
-      address: true,
-      bytecode: false,
-    });
+    const { address, abi } =
+      (await this.loader.load(query, {
+        abi: true,
+        address: true,
+        bytecode: false,
+      })) || {};
 
     if (typeof address !== 'string')
       throw new Error('Unable to parse contract address');
@@ -105,11 +106,33 @@ export default class EthersAdapter implements IAdapter {
       });
     }
   }
-  async getTransactionReceipt(transactionHash: string) {
-    return this.provider.getTransactionReceipt(transactionHash);
+  async _getTransactionReceipt(transactionHash: string) {
+    const receipt = await this.provider.getTransactionReceipt(transactionHash);
+    if (receipt == null)
+      throw new Error(
+        `Transaction receipt not found (transaction: ${transactionHash})`,
+      );
+    return receipt;
   }
   async waitForTransaction(transactionHash: string) {
     return this.provider.waitForTransaction(transactionHash);
+  }
+  async getTransactionReceipt(transactionHash: string) {
+    let receipt;
+    try {
+      // Attempt to get the receipt immediately; the transaction may have
+      // already been mined, or we're running on TestRPC with no mining time.
+      receipt = await this._getTransactionReceipt(transactionHash);
+    } catch (error) {
+      // Ignore the error if the receipt wasn't found
+      if (!error.toString().includes('Transaction receipt not found'))
+        throw error;
+    }
+
+    // Wait until the transaction has been mined, then get the receipt.
+    await this.waitForTransaction(transactionHash);
+    receipt = this._getTransactionReceipt(transactionHash);
+    return receipt;
   }
   /**
    * Sign a message hash (as binary) and return a split signature.